requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==1.26.4
biopython==1.83
torch==2.2.2 --index-url https://download.pytorch.org/whl/cpu
python-multipart